+++ changelog.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Added generic support for Liferay 7.3
 - Bugfix: Bugfix renaming packages inside bnd.bnd files
 - Bugfix: Detecting the Liferay version not working correctly in submodules
+- Bugfix: Resolving files relative to the content root directories in XML files using file references
 
 ## 0.0.8
 - Bugfix: Bugfix detecting a wrong version number in the Bundle-Version instruction inside the bnd.bnd file.